alfie-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
# Add this to backend/requirements.txt
gradientai==1.13.1